feat: implement JS interop layer and event dispatcher for Spotify Web Playback SDK
